requirements.txt
Flask==2.2.1
keras==2.10.0
numpy==1.23.2
Pillow==9.3.0
tensorflow==2.10.0
Werkzeug==2.2.2
ultralytics
